requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
future==0.18.2
idna==2.8
mypy==0.761
mypy-extensions==0.4.3
pycparser==2.19
python-dotenv==0.11.0
python-telegram-bot==12.4.2
requests==2.22.0
six==1.14.0
tornado==6.0.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
